usb_desc_dump: build form arduino ino file
